requirements.txt
# 필수 설치 

langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
streamlit-chat
gTTS

# 터미널 = pip install -r requirements.txt